working on integrating virtcap settings in sysfs

+++ software/firmware/pru0-shepherd-fw/include/lookup_table.h
@@ -6,7 +6,7 @@
 extern "C" {
 #endif
 
-int32_t lookup(int16_t table[][9], int32_t current, int32_t dbg[]);
+int32_t lookup(int32_t table[][9], int32_t current, int32_t dbg[]);
 void lookup_init();
 
 #ifdef __cplusplus
